requirements.txt
fastapi~=0.115.2
pydantic~=2.9.2
langchain~=0.3.3
chromadb~=0.5.13
langchain-community
langchain-openai
langchain-chroma

uvicorn[standard]
gunicorn
azure-storage-blob